% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{read.egonet}
\alias{read.egonet}
\title{Reads egocentric graph in table format and creates a matrix suitable for SNA.}
\usage{
read.egonet(file, sep = ",", dec = ".", ego.name = "EGO")
}
\arguments{
\item{file}{the name of the file which the data are to be read from. It must be a square matrix, with node's names in the first row and column. See also file of read.table}

\item{sep}{as in read.table}

\item{dec}{as in read.table}

\item{ego.name}{column and row name given to the ego subject. The default is \code{EGO}}
}
\value{
A square matrix containing the ego-centric network of the data in the file.

Per tutte le altre funzioni, l'output e' un \code{vector} della stessa
lunghezza di \code{testo} ma con testi normalizzati.
}
\description{
Reads egocentric graph in table format and creates a matrix suitable for SNA
Reads text table and creates ego-centric graph from it, with the ego actor in the first row and column. The result is a square matrix with common input/output connection' names.
}
\note{
%% ~~further notes~~
}
\examples{

dat <- read.egonet(system.file("extdata", "q1.csv", package="ARI"))
dat

}
\author{
Livio Finos, Andrea Sciandra
}
\keyword{~kwd1}
\keyword{~kwd2}
